1.7.2-1.1.0
1.1.0:
- Added 5 new mobs:
Human - It wanders around the world and randomly spawns with armor. This
mob has a very rare chance of dropping the current armor it is wearing.
Human Archer - Attacks mobs and slimes with a bow. Spawns with either a
Bow or a Ruby Bow, and it randomly spawns with armor. This mob drops the
current bow it is holding. This mob has a very rare chance of dropping
the current armor it is wearing.
Human Knight - Attacks mobs and slimes  with a sword. Spawns with any of
the swords (including the Ruby sword), and it randomly spawns with
armor. This mob drops the current sword it is holding.
Human Archer (Evil) - Like the Human Archer, but it targets the player,
other human mobs, and villagers instead of mobs and slimes.
Human Knight (Evil) - Like the Human Knight, but it targets the player,
other human mobs, and villagers instead of mobs and slimes.
- Fixed a bug where the Ruby Sword's damage would be the fist's damage.
- Added a new item: Primed TNT Spawner: This item is a Creative
Mode-only item and it allows you to spawn Primed TNT. This item is
useful for traps or just wanting to blow stuff up faster than normal
TNT. This item can be used in Dispensers.
- Both the Primed TNT Spawner and the Ender Crystal Spawner will now
work like Spawn Eggs in Dispensers.
- Added a new item: Medkit: Almost out of health and you don't have
anything that can heal you? Try our new Medkit! This handy Medkit will
heal 3.5 hearts so you can finally beat up that Zombie trying to bring
you down! Make sure you have plenty!